move alert-rules to monitoring helpers
